Remove publicApps util
